+++ src/interrogate/config.py
@@ -38,6 +38,7 @@ class InterrogateConfig:
     ignore_semiprivate = attr.ib(default=False)
     ignore_init_method = attr.ib(default=False)
     ignore_init_module = attr.ib(default=False)
+    include_regex = attr.ib(default=False)
 
 
 def find_project_root(srcs):
@@ -127,6 +128,14 @@ def read_pyproject_toml(ctx, param, value):
 
     if ctx.default_map is None:
         ctx.default_map = {}
+
+    # for backwards compatibility. before 1.1.3, only one regex was allowed.
+    # with 1.1.3+, multiple regexes can be provided, but we want to honor
+    # those that configured their pyproject.toml to be a single regex
+    # string (since now we're expecting a list of strings).
+    if "ignore_regex" in config:
+        if isinstance(config["ignore_regex"], str):
+            config["ignore_regex"] = [config["ignore_regex"]]
 
     ctx.default_map.update(config)
     return value